+++ graph.h
@@ -14,13 +14,13 @@ typedef struct NodeStruct Location;
 typedef struct Vector Node;
 
 Graph *GRAPHinit(int v, int a);
-void GRAPHinsertE(Graph *g, int index, int adj, float custo);
+void GRAPHinsertE(Graph *g, int index, int adj, double custo);
 void GRAPHremoveE(Graph *g, int i, int index);
 void GRAPHdestroy(Graph *g);
 void GRAPHaddV(Graph *g, int localidade, char *interesse);
 
 int modoA0(Graph *g, int v);
-float modoB0(Graph *g,int edge1,int edge2);
+double modoB0(Graph *g,int edge1,int edge2);
 int modoC0 (Graph *g, int v, int k);
 int vizinho(Graph *g, int v, int *visited, int maxstage, int stage);
 int modoD0(Graph *g, int v, int k);
